feat!: Add qubit hints on Guppy functions, allowing elision when building emulators (#1378)
Added a keyword argument to `@guppy` decorator for annotation of maximum
qubits required for the function, which allows the emulator to be built
without specifying a number of qubits (if given on the entrypoint
function):
```python
@guppy(max_qubits=1)
def main() -> None:
    result("c", measure(qubit()))

main.emulator().run()
```
The feature (still) complains if you:
- pass no qubit count anywhere
- pass a qubit count to the emulator that is insufficient to cover the
qubit count hinted on the entrypoint

The emulation will still crash if you do sufficient hinting, but the
allocated number of qubits is too low for the number of required qubits.

Additionally (or perhaps more importantly for future developments), this
PR adds a notion of metadata to the Guppy compiler which gets passed
onto the HUGR. This metadata is currently only attached to function
definitions, but could be extended to live on all definitions (and thus
be able to be attached to all HUGR nodes that are created from Guppy).
For a full list of TODOs, see below.

BREAKING CHANGE:
- The first argument to `add_unitarity_metadata` is now named `node`
instead of `func`, since its type was raised to allow for more HUGR
nodes to be fed. Migration is trivial.

> [!NOTE]
> This PR is only "breaking" `guppylang-internals`, and not `guppylang`

Before this PR is merged (or relatively soon thereafter), the following
things must/should/will happen:
- HUGR adds type-safe metadata access and refactors its core keys:
https://github.com/Quantinuum/hugr/pull/2755
- HUGR exports type-safe access and core keys to Python bindings:
https://github.com/Quantinuum/hugr/issues/2757
- Guppy uses the new core keys internally
- TKET uses the new pattern to define a unitary key
(e.g.`tket.modify.unitarity`, replacing the [string
literal](https://github.com/Quantinuum/tket2/blob/2918a67c6c5ce688ccd184192284355205047091/tket/src/modifier.rs#L66))
and defines a new `max_qubits` key (e.g. `tket.hint.max_qubits`),
exporting both to Python bindings
- Guppy adds TKET as a direct dependency
- Guppy uses the new keys internally to replace `"unitary"` as well as
the existing string literal for MaxQubits
- Guppy adds the ability to propagate metadata to any HUGR node,
attaching it to any def (?)

Closes #1297